+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2019-12-23
+### Add
+- RBAC - Migration
+- RBAC - Seeder
+- RBAC - APIs
+- RBAC - Middleware (RolePermission)
+
+### Fixed
+- Add Domain to Group API Bug
+- Operation Log Save IP Bug
+
 ## [0.8.2] - 2019-12-09
 ### Fixed
 - Batch Domain & CDN Logic